selects colors

+++ AnimatingArt2/Structs.pde
@@ -1,10 +1,10 @@
 class Pixel {
-  public float x;
-  public float y;
+  public int x;
+  public int y;
   public color c;
   public int shape;
 
-  Pixel (float x_, float y_, color c_) {
+  Pixel (int x_, int y_, color c_) {
     x = x_;
     y = y_;
     c = c_;
@@ -12,16 +12,17 @@ class Pixel {
   
   public void move () {
     Direction direction = shapes.get(this.shape);
+    //println(this.x + " " +direction.x);
     this.x += direction.x;
     this.y += direction.y;
   }
 }
 
 class Direction {
-  public float x;
-  public float y;
+  public int x;
+  public int y;
   
-  Direction (float x_, float y_) {
+  Direction (int x_, int y_) {
     x = x_;
     y = y_;
   }
